add: schedule detail, recylebin data film

+++ app/Http/Controllers/CinemaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Cinema;
+use App\Models\Schedule;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Exports\CinemaExport;
@@ -105,8 +106,12 @@ class CinemaController extends Controller
      */
     public function destroy($id)
     {
+        $schedules = Schedule::where('cinema_id', $id)->count();
+        if ($schedules) {
+            return redirect()->route('admin.cinemas.index')->with('error', 'Gagal! tidak dapat menghapus data bioskop! Data tertaut dengan jadwal tayang');
+        }
         Cinema::where('id', $id)->delete();
-        return redirect()->route('admin.cinemas.index')->with('Success', 'Berhasil menghapus data!');
+        return redirect()->route('admin.cinemas.index')->with('success', 'Berhasil menghapus data!');
     }
 
     public function export()
